[1] MAJOR RELEASE CHANGE. [2] -Wall (print all warnings) now standard in build. [3] Corrected all warnings. [3.1] #1 warning: unused variables. [3.2] #2 warning: signed vs. unsigned character comparison. [3.3] #3 warning: comparing values from two different enums (legit). [4] Several DAY ONE BUGS fixed! [5] Sanity, functional, valgrind unit test suites on Intel i7 PASSED. [6] WARNING: some functional tests fixed but not yet run.

+++ Hazer/inc/com/diag/hazer/datagram.h
@@ -55,11 +55,11 @@ typedef struct DatagramHeader {
  */
 enum {
     DATAGRAM_SIZE = (
-        (HAZER_NMEA_LONGEST > YODEL_UBX_LONGEST)
-            ? ((HAZER_NMEA_LONGEST > TUMBLEWEED_RTCM_LONGEST)
+        ((size_t)HAZER_NMEA_LONGEST > (size_t)YODEL_UBX_LONGEST)
+            ? (((size_t)HAZER_NMEA_LONGEST > (size_t)TUMBLEWEED_RTCM_LONGEST)
                 ? HAZER_NMEA_LONGEST
                 : TUMBLEWEED_RTCM_LONGEST)
-            : ((YODEL_UBX_LONGEST > TUMBLEWEED_RTCM_LONGEST)
+            : (((size_t)YODEL_UBX_LONGEST > (size_t)TUMBLEWEED_RTCM_LONGEST)
                 ? YODEL_UBX_LONGEST
                 : TUMBLEWEED_RTCM_LONGEST)
     ),
